Add a path for barcode rescue in parsing
Due to read length and presence of some homopolymers in the
linker sequence, it is not uncommon that L1 and L2 align with
too few basepairs between them (I have yet to see a case where
this interval is too large). In such a case, if the L2 alignment
is higher than the L1 alignment, shift L1 appropriately

It is possible that L2 might need to be shifted, or *both* L1 and L2,
but I have not yet observed such cases.