Bug fix: query name too long
Previously: If the first linker started within 1 sample barcode length of the
UMI, a negative offset would result in the 'raw umi' value being nearly the
full length of the read; with the 'parsed umi' value properly set to '*'. This
can lead, downstream, to a 'query name too long' failure during bwa.

Current: If the first linker starts within 1 sample barcode length of the read
start, the negative offset is clamped to 0, so the barcode is correctly set to
NNNNNNNNNN, shortening the final read name.